refactor: split large dev helpers and config tables
