+++ src/config/explosions/alliance.ts
@@ -1,0 +1,134 @@
+import {
+  DEFAULT_PARTICLE_COUNTS,
+  DEFAULT_SHOCKWAVE_MULTI,
+  DEFAULT_TIMING,
+  EXPLOSION_SCALE,
+  ExplosionPresetMap,
+  scaleRange,
+} from './common.js';
+
+export const ALLIANCE_EXPLOSIONS: ExplosionPresetMap = {
+  fighter: {
+    baseRadius: 3 * EXPLOSION_SCALE,
+    flashIntensity: 1.1,
+    lightColor: '#a6d8ff',
+    lightFalloff: 40,
+    debrisCount: 6 * EXPLOSION_SCALE,
+    particleCounts: { ...DEFAULT_PARTICLE_COUNTS, sparks: 12 * EXPLOSION_SCALE },
+    palette: {
+      flash: '#a6d8ff',
+      shockwave: '#9fc9ff',
+      fireballHot: '#f0fbff',
+      smoke: '#6b7b92',
+    },
+    timing: {
+      duration: 1.5,
+      lightDuration: 0.2,
+      shockwave: { delay: 0.06, duration: 0.28 },
+      fireball: { delay: 0.18, duration: 0.35 },
+      debrisSpeed: scaleRange([10, 22]),
+    },
+    shockwaveMaxRadiusMulti: DEFAULT_SHOCKWAVE_MULTI,
+  },
+  corvette: {
+    baseRadius: 3.6 * EXPLOSION_SCALE,
+    flashIntensity: 1.15,
+    lightColor: '#a6d8ff',
+    lightFalloff: 55,
+    debrisCount: 8 * EXPLOSION_SCALE,
+    particleCounts: {
+      ...DEFAULT_PARTICLE_COUNTS,
+      sparks: 14 * EXPLOSION_SCALE,
+      plasma: 8 * EXPLOSION_SCALE,
+    },
+    palette: {
+      flash: '#a6d8ff',
+      shockwave: '#8fb6ff',
+      fireballHot: '#e4f6ff',
+      smoke: '#65728d',
+    },
+    timing: {
+      duration: 1.65,
+      lightDuration: 0.22,
+      shockwave: { delay: 0.07, duration: 0.3 },
+      fireball: { delay: 0.19, duration: 0.38 },
+      debrisSpeed: scaleRange([11, 24]),
+    },
+    shockwaveMaxRadiusMulti: DEFAULT_SHOCKWAVE_MULTI,
+  },
+  frigate: {
+    baseRadius: 4.4 * EXPLOSION_SCALE,
+    flashIntensity: 1.2,
+    lightColor: '#a6d8ff',
+    lightFalloff: 65,
+    debrisCount: 10 * EXPLOSION_SCALE,
+    particleCounts: {
+      sparks: 16 * EXPLOSION_SCALE,
+      plasma: 10 * EXPLOSION_SCALE,
+      smoke: 12 * EXPLOSION_SCALE,
+    },
+    palette: {
+      flash: '#a6d8ff',
+      shockwave: '#85acff',
+      fireballHot: '#d7ecff',
+      smoke: '#5f6c87',
+    },
+    timing: {
+      ...DEFAULT_TIMING,
+      debrisSpeed: scaleRange(DEFAULT_TIMING.debrisSpeed),
+    },
+    shockwaveMaxRadiusMulti: DEFAULT_SHOCKWAVE_MULTI,
+  },
+  destroyer: {
+    baseRadius: 5.4 * EXPLOSION_SCALE,
+    flashIntensity: 1.25,
+    lightColor: '#a6d8ff',
+    lightFalloff: 75,
+    debrisCount: 12 * EXPLOSION_SCALE,
+    particleCounts: {
+      sparks: 18 * EXPLOSION_SCALE,
+      plasma: 12 * EXPLOSION_SCALE,
+      smoke: 14 * EXPLOSION_SCALE,
+    },
+    palette: {
+      flash: '#a6d8ff',
+      shockwave: '#7ca3ff',
+      fireballHot: '#cbe1ff',
+      smoke: '#59647f',
+    },
+    timing: {
+      duration: 2.0,
+      lightDuration: 0.28,
+      shockwave: { delay: 0.09, duration: 0.36 },
+      fireball: { delay: 0.22, duration: 0.45 },
+      debrisSpeed: scaleRange([13, 28]),
+    },
+    shockwaveMaxRadiusMulti: DEFAULT_SHOCKWAVE_MULTI,
+  },
+  carrier: {
+    baseRadius: 7 * EXPLOSION_SCALE,
+    flashIntensity: 1.3,
+    lightColor: '#a6d8ff',
+    lightFalloff: 90,
+    debrisCount: 16 * EXPLOSION_SCALE,
+    particleCounts: {
+      sparks: 20 * EXPLOSION_SCALE,
+      plasma: 14 * EXPLOSION_SCALE,
+      smoke: 18 * EXPLOSION_SCALE,
+    },
+    palette: {
+      flash: '#a6d8ff',
+      shockwave: '#7298ff',
+      fireballHot: '#bed4ff',
+      smoke: '#525c77',
+    },
+    timing: {
+      duration: 2.2,
+      lightDuration: 0.32,
+      shockwave: { delay: 0.1, duration: 0.4 },
+      fireball: { delay: 0.25, duration: 0.5 },
+      debrisSpeed: scaleRange([15, 32]),
+    },
+    shockwaveMaxRadiusMulti: DEFAULT_SHOCKWAVE_MULTI,
+  },
+};
