+++ src/config/explosions/reavers.ts
@@ -1,0 +1,139 @@
+import {
+  DEFAULT_SHOCKWAVE_MULTI,
+  EXPLOSION_SCALE,
+  ExplosionPresetMap,
+  scaleRange,
+} from './common.js';
+
+export const REAVERS_EXPLOSIONS: ExplosionPresetMap = {
+  fighter: {
+    baseRadius: 3.2 * EXPLOSION_SCALE,
+    flashIntensity: 1.2,
+    lightColor: '#ff8447',
+    lightFalloff: 40,
+    debrisCount: 7 * EXPLOSION_SCALE,
+    particleCounts: {
+      sparks: 14 * EXPLOSION_SCALE,
+      plasma: 6 * EXPLOSION_SCALE,
+      smoke: 9 * EXPLOSION_SCALE,
+    },
+    palette: {
+      flash: '#ffb347',
+      shockwave: '#ff7847',
+      fireballHot: '#ff6138',
+      smoke: '#5a3122',
+    },
+    timing: {
+      duration: 1.6,
+      lightDuration: 0.22,
+      shockwave: { delay: 0.06, duration: 0.3 },
+      fireball: { delay: 0.17, duration: 0.38 },
+      debrisSpeed: scaleRange([12, 25]),
+    },
+    shockwaveMaxRadiusMulti: DEFAULT_SHOCKWAVE_MULTI,
+  },
+  corvette: {
+    baseRadius: 3.9 * EXPLOSION_SCALE,
+    flashIntensity: 1.25,
+    lightColor: '#ff8447',
+    lightFalloff: 55,
+    debrisCount: 9 * EXPLOSION_SCALE,
+    particleCounts: {
+      sparks: 16 * EXPLOSION_SCALE,
+      plasma: 8 * EXPLOSION_SCALE,
+      smoke: 11 * EXPLOSION_SCALE,
+    },
+    palette: {
+      flash: '#ffb347',
+      shockwave: '#ff6a36',
+      fireballHot: '#ff4c24',
+      smoke: '#53291c',
+    },
+    timing: {
+      duration: 1.75,
+      lightDuration: 0.25,
+      shockwave: { delay: 0.07, duration: 0.32 },
+      fireball: { delay: 0.18, duration: 0.4 },
+      debrisSpeed: scaleRange([13, 26]),
+    },
+    shockwaveMaxRadiusMulti: DEFAULT_SHOCKWAVE_MULTI,
+  },
+  frigate: {
+    baseRadius: 4.8 * EXPLOSION_SCALE,
+    flashIntensity: 1.32,
+    lightColor: '#ff8447',
+    lightFalloff: 65,
+    debrisCount: 12 * EXPLOSION_SCALE,
+    particleCounts: {
+      sparks: 18 * EXPLOSION_SCALE,
+      plasma: 10 * EXPLOSION_SCALE,
+      smoke: 14 * EXPLOSION_SCALE,
+    },
+    palette: {
+      flash: '#ffb347',
+      shockwave: '#ff612c',
+      fireballHot: '#ff3d18',
+      smoke: '#4a2216',
+    },
+    timing: {
+      duration: 1.9,
+      lightDuration: 0.27,
+      shockwave: { delay: 0.08, duration: 0.34 },
+      fireball: { delay: 0.19, duration: 0.42 },
+      debrisSpeed: scaleRange([14, 28]),
+    },
+    shockwaveMaxRadiusMulti: DEFAULT_SHOCKWAVE_MULTI,
+  },
+  destroyer: {
+    baseRadius: 5.8 * EXPLOSION_SCALE,
+    flashIntensity: 1.38,
+    lightColor: '#ff8447',
+    lightFalloff: 80,
+    debrisCount: 15 * EXPLOSION_SCALE,
+    particleCounts: {
+      sparks: 22 * EXPLOSION_SCALE,
+      plasma: 12 * EXPLOSION_SCALE,
+      smoke: 16 * EXPLOSION_SCALE,
+    },
+    palette: {
+      flash: '#ffb347',
+      shockwave: '#ff531d',
+      fireballHot: '#ff2f0e',
+      smoke: '#421b11',
+    },
+    timing: {
+      duration: 2.1,
+      lightDuration: 0.3,
+      shockwave: { delay: 0.09, duration: 0.38 },
+      fireball: { delay: 0.21, duration: 0.48 },
+      debrisSpeed: scaleRange([15, 30]),
+    },
+    shockwaveMaxRadiusMulti: DEFAULT_SHOCKWAVE_MULTI,
+  },
+  carrier: {
+    baseRadius: 7.6 * EXPLOSION_SCALE,
+    flashIntensity: 1.45,
+    lightColor: '#ff8447',
+    lightFalloff: 95,
+    debrisCount: 20 * EXPLOSION_SCALE,
+    particleCounts: {
+      sparks: 26 * EXPLOSION_SCALE,
+      plasma: 16 * EXPLOSION_SCALE,
+      smoke: 20 * EXPLOSION_SCALE,
+    },
+    palette: {
+      flash: '#ffb347',
+      shockwave: '#ff4512',
+      fireballHot: '#ff2405',
+      smoke: '#38150c',
+    },
+    timing: {
+      duration: 2.4,
+      lightDuration: 0.35,
+      shockwave: { delay: 0.1, duration: 0.42 },
+      fireball: { delay: 0.23, duration: 0.55 },
+      debrisSpeed: scaleRange([16, 35]),
+    },
+    shockwaveMaxRadiusMulti: DEFAULT_SHOCKWAVE_MULTI,
+  },
+};
